Move bindTestLogFlags to test.go
